+++ src/stratega/src/Representation/GameInfo.cpp
@@ -89,7 +89,7 @@ namespace SGA
 		std::unordered_map<int, int> counts;
 		for (auto techTree : technologyTreeCollection->technologyTreeTypes)
 		{
-			counts.emplace(techTree.first, techTree.second.technologies.size());
+			counts.emplace(techTree.first, (int)techTree.second.technologies.size());
 		}
 		/*for (const auto& [id, techTree] : technologyTreeCollection->technologyTreeTypes)
 		{
